Added: Started on functionality on /user/:user_id Changes: - All repositories except Status is changes from id of integer to long

+++ server/server/src/main/java/se/experis/tidsbanken/server/repositories/IneligiblePeriodRepository.java
@@ -3,6 +3,6 @@ package se.experis.tidsbanken.server.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.experis.tidsbanken.server.models.IneligiblePeriod;
 
-public interface IneligiblePeriodRepository extends JpaRepository<IneligiblePeriod, Integer> {
-    IneligiblePeriod getById(int id);
+public interface IneligiblePeriodRepository extends JpaRepository<IneligiblePeriod, Long> {
+    IneligiblePeriod getById(long id);
 }
