+++ server/server/src/main/java/se/experis/tidsbanken/server/repositories/VacationRequestRepository.java
@@ -3,6 +3,6 @@ package se.experis.tidsbanken.server.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.experis.tidsbanken.server.models.VacationRequest;
 
-public interface VacationRequestRepository extends JpaRepository<VacationRequest, Integer> {
-    VacationRequest getById(int id);
+public interface VacationRequestRepository extends JpaRepository<VacationRequest, Long> {
+    VacationRequest getById(long id);
 }
